base: added module support
Added module support for the bot. Modules are located in the "modules"
directory and have to have a class within that is the same name as the
module file (without the .py) For example:

    ######################
    # modules/TestMod.py #
    ######################
    class TestMod(object):
        pass

the __init__ function takes in an instance of the bot. Used for
accessing server information and message sending functions. Within the
__init__ function, self.commands has to be initialized similar to the
bot's self.commands attribute. All the functions referenced in the
commands map must also be @asyncio.coroutine functions. Modules must
also have a unload function (note: this is not an @asyncio.coroutine
function). This function is used to free any resources used by the bot
before unloading. This function may be empty.

Modules are loaded and unloaded with !load and !unload respectively.